% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap}
\alias{bootstrap}
\title{Bootstrap a Data Frame}
\usage{
bootstrap(data)
}
\arguments{
\item{data}{Data frame}
}
\value{
Data frame
}
\description{
Form a bootstrap sample from a data frame \code{data} by sampling
\code{nrow(data)} rows of the data frame with replacement.
}
