% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_csv.R
\name{cat_csv.gz}
\alias{cat_csv.gz}
\title{Concatenate CSVs inside a .tar.gz}
\usage{
cat_csv.gz(gzfile, pattern, csv_folder = ".")
}
\arguments{
\item{gzfile}{Archive}

\item{pattern}{Pattern which the CSV files are to match}

\item{csv_folder}{Folder within the archive in which the CSVs are found}
}
\value{
Data frame
}
\description{
Row-bind several CSV files that live inside a tar.gz archive. It is
assumed that these CSV files are contained within a directory called
\code{foo} where \code{foo.tar.gz} is the filename of the archive; if
this is not the case, unexpected behaviour will occur.
}
\details{
The \code{csv_folder} argument can be used to specify an optional directory 
within the \code{foo} directory that contains the CSV files.
}
