% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_csv.R
\name{cat_csv}
\alias{cat_csv}
\title{Concatenate CSVs}
\usage{
cat_csv(folder, pattern)
}
\arguments{
\item{folder}{Folder in which the CSV files live}

\item{pattern}{Pattern which included files are to match}
}
\value{
A data frame
}
\description{
Row-bind several CSV files into one dataframe.
}
