% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closest_in_vector.R
\name{closest_in_vector}
\alias{closest_in_vector}
\title{Get the Element(s) of a Vector Closest to a Reference Value}
\usage{
closest_in_vector(x, vec, k = 1)
}
\arguments{
\item{x}{Reference value}

\item{vec}{Vector}

\item{k}{How many closest neighbours to return}
}
\value{
A data frame with the following columns, sorted by increasing distance:
\describe{
\item{\code{index}}{Index of element (i.e. \code{value} is the same as \code{vec[index]})}
\item{\code{value}}{Element}
\item{\code{dist}}{Euclidean distance of element to reference value}
}
}
\description{
Return the \emph{k} elements of a vector closest (in terms of Euclidean distance) to a reference value, as well as those elements' indices in the vector.
}
