% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df2tex.R
\name{df2tex}
\alias{df2tex}
\title{Data Frame to TeX Tabular}
\usage{
df2tex(df, file, col.names = TRUE, row.names = TRUE, header = "",
  before = NULL, after = NULL)
}
\arguments{
\item{df}{Data frame}

\item{file}{File to write the table into}

\item{col.names}{Should column names be printed?}

\item{row.names}{Should row names be printed?}

\item{header}{Optional tabular header. If not specified, a default header of the form \code{\\tabular{r...r}} is used with as many columns as there are columns in the data frame.}

\item{before}{Optional vector of lines to write after header and (possible) column names but before data. Mostly useful to add custom column names, or an \code{\\hline}.}

\item{after}{Optional vector of lines to write before \code{\\end{tabular}}}
}
\description{
Prints a data frame as a (La)TeX tabular.
}
