% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor2character.R
\name{factor2character}
\alias{factor2character}
\title{Factor to Character}
\usage{
factor2character(x)
}
\arguments{
\item{x}{Factor to be transformed}
}
\value{
Character vector, i.e. the factor without the levels.
}
\description{
Turns a factor into a plain character vector.
}
