% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor2numeric.R
\name{factor2numeric}
\alias{factor2numeric}
\title{Factor to Numeric}
\usage{
factor2numeric(x)
}
\arguments{
\item{x}{Factor to be transformed}
}
\value{
Numeric vector
}
\description{
Turns a factor into a numeric vector.
}
