% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hipster.R
\name{get_hipster}
\alias{get_hipster}
\title{get_hipster}
\usage{
get_hipster()
}
\description{
Plots an image of a hipster.  What else?
}
\examples{
# Get a hipster
get_hipster()
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
