% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knockout_lm.R
\name{knockout_lm}
\alias{knockout_lm}
\title{Knock Out Outliers in a Linear Regression}
\usage{
knockout_lm(data, formula, id.var)
}
\arguments{
\item{data}{Data frame}

\item{formula}{Regression formula}

\item{id.var}{Name of identifier variable (column)}
}
\value{
Data frame with the following columns:
\describe{
\item{\code{iteration}}{Knockout iteration}
\item{\code{knockee}}{Data pointed knocked out at this iteration}
\item{\code{RSS_reduction}}{Reduction in residual sum of squares resulting from the knockout}
}
}
\description{
Knock out outliers in a linear regression by recursively pruning those
data points that contribute most to the regression error, operationalized
as the residual sum of squares at each knockout iteration.
}
