% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic.R
\name{logistic}
\alias{logistic}
\title{Logistic Function}
\usage{
logistic(t, s = 1, k = 0, U = 1, L = 0)
}
\arguments{
\item{t}{Variable (usually time, hence \code{t})}

\item{s}{Slope}

\item{k}{Intercept}

\item{U}{Upper asymptote}

\item{L}{Lower asymptote}
}
\value{
Value of the function at \code{t}, given the parameters.
}
\description{
Generalized logistic function.
}
