% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logseq.R
\name{logseq}
\alias{logseq}
\title{Logarithmic Sequences}
\usage{
logseq(from, to, length.out)
}
\arguments{
\item{from}{Starting point of sequence}

\item{to}{Endpoint of sequence}

\item{length.out}{Desired length of sequence}
}
\value{
A numeric vector
}
\description{
Generates a logarithmically spaced sequence.
}
