% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prettyround.R
\name{prettyround}
\alias{prettyround}
\title{Pretty Print Rounded Figures}
\usage{
prettyround(x, digits)
}
\arguments{
\item{x}{Number}

\item{digits}{Number of digits}
}
\value{
Pretty-printed number. Note that this is of class character, not numeric.
}
\description{
Pretty print a rounded figure so that all trailing digits are shown,
including zeroes.
}
