% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seed_seed.R
\name{seed_seed}
\alias{seed_seed}
\title{Seed Seed}
\usage{
seed_seed(n, seed, ...)
}
\arguments{
\item{n}{Number (e.g. id of parallel job)}

\item{seed}{Initial seed; this should be set so that results are replicable}

\item{...}{Further parameters passed to \code{set.seed}}
}
\value{
The final seed, i.e. the \code{n}th number in the sequence
}
\description{
First, seeds the random number generator with \code{seed}. Then generates
a sequence of \code{n} of random integers between 1 and
\code{.Machine$integer.max}. Finally, seeds the random number generator
with the \code{n}th (i.e. last) number in this sequence. The point of this
function is that it may be used to seed the RNG at the start of a job
that is run on a cluster, so that parallel jobs are using - if not
statistically independent random number streams - at least not the exact
same stream.
}
