% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_csv.R
\name{cat_csv.gz}
\alias{cat_csv.gz}
\title{Concatenate CSVs inside a .tar.gz}
\usage{
cat_csv.gz(gzfile, pattern, csv_folder = ".")
}
\arguments{
\item{gzfile}{Archive}

\item{pattern}{Pattern which the CSV files are to match}

\item{csv_folder}{Folder within the archive in which the CSVs are found}
}
\value{
Data frame
}
\description{
Row-bind several CSV files that live inside a tar.gz archive. The
\code{csv_folder} argument specifies an optional directory within the
archive that contains the CSV files.
}
