% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refresh.R
\name{refresh}
\alias{refresh}
\title{Package Refresh}
\usage{
refresh(pkg)
}
\arguments{
\item{pkg}{Package name (character vector)}
}
\description{
"Refresh" a package by unloading its namespace and immediately re-requiring
that namespace. This is useful if an updated version of the package has been
installed by another process on the machine in the meantime, to avoid
having to restart the R session.
}
